% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{byCountry_}
\alias{byCountry_}
\title{byCountry_}
\format{
An object of class \code{list} of length 4.
}
\usage{
byCountry_
}
\description{
foRgotten::FE(CE = data_byCountry, maxOrder = 2, thr = drr[i], reps = 1000)
}
\keyword{datasets}
