% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{byYear}
\alias{byYear}
\title{byYear}
\format{
An object of class \code{list} of length 6.
}
\usage{
byYear
}
\description{
foRgotten::FE(CE = data_byYear, maxOrder = 2, thr = k[i], reps = 100)
}
\keyword{datasets}
