% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_By_Country_FE}
\alias{max_By_Country_FE}
\title{max_By_Country_FE usando foRgotten}
\format{
An object of class \code{list} of length 2.
}
\usage{
Ver ejemplo
}
\description{
Los datos fueron calculados usando la funcion FE de foRgotten.
Los parametros son: thr = 0.05, reps = 1000
}
\examples{
library(datosPxP)
library(foRgotten)
dd <- datosPxP::datos
max_By_Country <- max_ByCountry(dd)
summary(max_By_Country)
max_By_Country_FE <- foRgotten::FE(CE = max_By_Country, thr = 0.05, maxOrder = 2, reps = 1000)
max_By_Country_FE$boot$Order_2
}
\keyword{datasets}
