% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_By_Year_FE}
\alias{max_By_Year_FE}
\title{max_ByCountry usando foRgotten}
\format{
An object of class \code{list} of length 2.
}
\usage{
Ver ejemplo
}
\description{
Los datos fueron calculados usando la funcion FE de foRgotten.
Los parametros son: thr = 0.5, reps = 1000
}
\examples{
library(datosPxP)
library(foRgotten)
dd <- datosPxP::datos
max_By_Year <- maxByYear(dd)
summary(max_By_Year)
max_By_Year_FE <- foRgotten::FE(CE = max_By_Year, thr = 0.5, maxOrder = 2, reps = 1000)
head(max_By_Year_FE$boot$Order_2)
}
\keyword{datasets}
